[1548] Migration of pipeline component to use postresql (#1570)
* Add new database abstraction in common

* fly - migrate observations to postgres

* fly - migrate vaas and governor status to postgres

* fly - add event dispatcher for vaa-pipeline

* Add to repository on conflict behaviour to queries

* Add observations and vaas metrics to repository

* add support to vaa pythnet table

* Add new postgres repository for guardian set in common

* add common method to exec and scan

* add query log

* fly - add event dispatacher for governor config message

* update migration.sql file

* fix inserted flag issue, add metrics and event dispatcher for vaas and governor config

* Add comment to pgx cmdTag bug and workaround for this cases

* update pipeline vaa event & migration

* Add in fly-event-processor support to governor config chain

* Add in fly-event-processor governor_status postgres repository

* Fix table wh_governor_config_chains

* fix governor config chain flow

* fly-event-processor - fix governor config flow

* Add health for postgres in common

* add dummy governor status processor and refactor run and build

* Add comment in fly-event-processor

* migration of pipeline component

* finish full test

* fix broken merge

* add missing implementation

* small changes

* change condition

* add unit-test to pipeline

* add missing Close()

* fix unit-test

---------

+++ pipeline/pipeline/txhash_handler.go
@@ -86,7 +86,7 @@ func (t *TxHashHandler) Run(ctx context.Context) {
 						t.pushFunc(ctx, &item.Event)
 						delete(t.fixItems, vaaID)
 						// increment metrics vaa with txhash fixed
-						t.metrics.IncVaaWithTxHashFixed(item.Event.ChainID)
+						t.metrics.IncVaaWithTxHashFixed(uint16(item.Event.ChainID))
 					}
 				} else {
 					t.logger.Error("Vaa txhash fix failed", zap.String("vaaID", vaaID))
